refactor: move course related file to its own file
